conf/requirements.txt
aa-buybackprogram==2.3.3
aa-charlink==1.11.1
aa-fleetpings==3.9.0
aa-freight==2.0.0
aa-inactivity==2.0.1
aa-killtracker==1.2.0
aa-memberaudit-securegroups==1.0.1
aa-package-monitor==2.3.0
aa-ravworks-exporter==0.5.1
aa-srp==3.2.0
aa-statistics==0.2.1
aa-structures==3.3.1
aa-structuretimers==2.1.0
aa-taskmonitor==0.23.1
aa-timezones==2.4.3
allianceauth-afat==4.2.0
allianceauth-corptools-moons==1.0.1
allianceauth-discordbot==4.1.0
allianceauth-wiki-js==0.1.1
django-environ==0.12.0
django-health-check==3.20.0
fittings==2.2.1
psutil==7.1.3
